modules/requirements.txt
os
numpy
ipfml
sklearn
Pillow